[29기 유재민] 검색페이지 리팩토링
