+++ src/config.js
@@ -1,4 +1,4 @@
-export const BASE_URL = 'http://10.58.3.139:3000';
+export const BASE_URL = 'http://10.58.5.57:3000';
 
 export const api = {
   login: BASE_URL + '/users/callback',
@@ -7,4 +7,6 @@ export const api = {
   search: BASE_URL + '/search',
   redirect_uri: 'http://localhost:3000/oauth/callback/kakao',
   kakao_token_request: BASE_URL + '/users/callback',
+  product_list: BASE_URL + '/products/lists',
+  product_like: BASE_URL + '/products/likes',
 };
