+++ src/pages/Search/Search.js
@@ -1,31 +1,19 @@
-import React from 'react';
-import styled from 'styled-components';
-import { useState, useEffect } from 'react';
-import { Link } from 'react-router-dom';
-import theme from '../../styles/theme';
+import React, { useState, useEffect } from 'react';
+import SearchProduct from './SearchProduct';
 import { api } from '../../config';
+import styled from 'styled-components';
+import theme from '../../styles/theme';
 
 const Search = () => {
   const [searchData, setSearchData] = useState([]);
 
   useEffect(() => {
-    fetch(api.search, {
+    fetch(api.product_list, {
       method: 'GET',
     })
       .then(res => res.json())
-      .then(data => {
-        setSearchData(data);
-      });
+      .then(data => setSearchData(data.results));
   }, []);
-
-  const AssmentComponent = ({ url, name }) => {
-    return (
-      <AssessmentIndividual>
-        <AssessmentIcon src={'/images/Search/' + url} />
-        <AssessmentText>{name}</AssessmentText>
-      </AssessmentIndividual>
-    );
-  };
 
   return (
     <Wrap>
@@ -47,7 +35,7 @@ const Search = () => {
           ))}
         </FillterWrap>
         <ProductListWrap>
-          {searchData.length === 0 ? (
+          {searchData[0]?.length === 0 ? (
             <NoneSearchWrap>
               <NoneSearchTitle>
                 <NoneSearchValue>검색어 value가 들어갈 자리</NoneSearchValue>에
@@ -61,43 +49,12 @@ const Search = () => {
             </NoneSearchWrap>
           ) : (
             <ProductWrap>
-              {searchData?.map(product => (
-                <Product key={product.id}>
-                  <ProductLinkWrap to={product.route_url}>
-                    <ProductTumWrap>
-                      <ProductTum src={product.img_url} />
-                      {product.coupon && (
-                        <CouponWrap>
-                          <CouponTxt>{product.coupon}</CouponTxt>
-                        </CouponWrap>
-                      )}
-                    </ProductTumWrap>
-                    <ClassInfoWrap>
-                      <UserId>{product.user_id}</UserId>
-                      <ClassTitle>{product.class_title}</ClassTitle>
-                      <AssessmentWrap>
-                        <AssmentComponent
-                          name={product.class_heart}
-                          url="icon_gray_heart.png"
-                        />
-                        <AssmentComponent
-                          name={product.class_like}
-                          url="icon_gray_like.png"
-                        />
-                      </AssessmentWrap>
-                      <PriceWrap>
-                        <Sale>{product.sale}</Sale>
-                        <Price>월 {product.price}원</Price>
-                        <Term>({product.term}개월)</Term>
-                      </PriceWrap>
-                    </ClassInfoWrap>
-                  </ProductLinkWrap>
-                  <LikeIcon src="/images/Search/icon_like_inactive.png" />
-                </Product>
+              {searchData[0]?.map(props => (
+                <SearchProduct key={props.id} props={props} />
               ))}
             </ProductWrap>
           )}
-          {searchData.length > 0 && (
+          {searchData[0]?.length > 0 && (
             <PagenationWrap>
               <PageButtonInactive>
                 <span>&#60;</span>
@@ -185,165 +142,6 @@ const ProductWrap = styled.ul`
   -webkit-box-pack: start;
 `;
 
-const Product = styled.li`
-  position: relative;
-  width: 25%;
-  min-height: 1px;
-  padding-right: 12px;
-  padding-left: 12px;
-  margin-bottom: 32px;
-  box-sizing: border-box;
-}
-`;
-
-const ProductLinkWrap = styled(Link)`
-  display: inline-block;
-  width: 276px;
-  height: 207px;
-`;
-
-const ProductTumWrap = styled.div`
-  display: inline-block;
-  position: relative;
-  width: 100%;
-  height: 100%;
-  border-radius: 5px;
-  overflow: hidden;
-  cursor: pointer;
-`;
-
-const ProductTum = styled.img`
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  transform: scale(1);
-  transition: transform 0.3s ease 0s, opacity 0.1s linear 0s !important;
-
-  &:hover{
-    transform: scale(1.1);
-  }
-}
-`;
-
-const LikeIcon = styled.img`
-  position: absolute;
-  top: 15px;
-  right: 25px;
-  width: 21px;
-  height: 18px;
-  object-fit: cover;
-  cursor: pointer;
-`;
-
-const CouponWrap = styled.div`
-  position: absolute;
-  top: 15px;
-  left: 15px;
-`;
-
-const CouponTxt = styled.span`
-  margin: 0px;
-  padding: 6px 8px;
-  background-color: ${theme.pinkred};
-  color: ${theme.white};
-  border-radius: 2px;
-  font-size: 11px;
-  font-weight: 600;
-  line-height: 16px;
-`;
-
-const ClassInfoWrap = styled.div`
-  display: block;
-  flex-direction: column;
-`;
-
-const UserId = styled.p`
-  margin: 8px 0;
-  padding: 0;
-  border: 0;
-  color: ${theme.realblack};
-  font-size: 0.6875rem;
-  font-weight: 700;
-  line-height: 0.875rem;
-`;
-
-const ClassTitle = styled.h2`
-  display: -webkit-box;
-  height: 40px;
-  margin: 4px 0px 8px;
-  color: ${theme.realblack};
-  font-weight: normal;
-  font-size: 14px;
-  line-height: 20px;
-  letter-spacing: -0.15px;
-  overflow: hidden;
-  word-break: keep-all;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-`;
-
-const AssessmentWrap = styled.div`
-  display: flex;
-  padding-bottom: 10px;
-  border-bottom: 1px solid #f8f8f8;
-`;
-
-const AssessmentIndividual = styled.div`
-  ${theme.flexCustom('center', 'flex-start', 'row')};
-  margin: 0px 8px 0px 0px;
-  font-weight: normal;
-  line-height: 16px;
-  letter-spacing: normal;
-  -webkit-box-align: center;
-`;
-
-const AssessmentIcon = styled.img`
-  width: 12px;
-  height: 12px;
-  object-fit: cover;
-`;
-
-const AssessmentText = styled.span`
-  margin-left: 5px;
-  color: ${theme.gray};
-  font-size: 11px;
-`;
-
-const PriceWrap = styled.div`
-  ${theme.flexCustom('center', 'flex-start', 'row')};
-  padding-top: 10px;
-`;
-
-const Sale = styled.span`
-  margin-right: 3px;
-  padding: 0;
-  border: 0;
-  color: ${theme.pinkred};
-  font-size: 0.8125rem;
-  line-height: 1.125rem;
-  font-weight: 700;
-`;
-
-const Price = styled.span`
-  margin-right: 3px;
-  padding: 0;
-  border: 0;
-  color: ${theme.black};
-  font-size: 0.8125rem;
-  line-height: 1.125rem;
-  font-weight: 700;
-`;
-
-const Term = styled.span`
-  margin: 0;
-  padding: 0;
-  border: 0;
-  color: ${theme.gray};
-  font-size: 0.6875rem;
-  line-height: 0.875rem;
-  font-weight: 400;
-`;
-
 const NoneSearchWrap = styled.div`
   ${theme.flexCustom('center', 'center', 'column')};
   width: 100%;
@@ -375,11 +173,12 @@ const NoneSearchCheck = styled.span`
 `;
 
 const PagenationWrap = styled.div`
-  ${theme.flexCustom('center', 'center', 'row')};
+${theme.flexCustom('center', 'center', 'row')};
   width: 100%;
   margin-top: 32px;
   -webkit-box-pack: center;
   -webkit-box-align: center;
+  display: none;
 }
 `;
 
@@ -423,7 +222,6 @@ const PageButtonInactive = styled.button`
   }
 `;
 
-// 상수 데이터 : 필터 대분류
 const FILTER_MAIN_CATEGORY = [
   {
     id: 1,
@@ -439,7 +237,6 @@ const FILTER_MAIN_CATEGORY = [
   },
 ];
 
-// 상수 데이터 : 필터 오픈 시기별
 const FILTER_OPEN_TIME = [
   {
     id: 1,
